// e2e/models/api/authentication/create-account.model.ts
export interface CreateAccountAPIModel {
  name: string;
  email: string;
  password: string;
  title: string;
  birth_date: string;
  birth_month: string;
  birth_year: string;
  firstname: string;
  lastname: string;
  company: string;
  address1: string;
  address2: string;
  country: string;
  zipcode: string;
  state: string;
  city: string;
  mobile_number: string;
}

export interface CreateAccountBodyAPIModel {
  responseCode: number;
  message: string;
}